requirements.txt
secrets
matplotlib
numpy
skimage
pillow